fixed detecting drop node
