+++ src/components/ui/TreeNode.vue
@@ -1,9 +1,8 @@
 <script setup lang="ts">
-import { type PropType, computed, ref } from 'vue'
+import { type PropType, computed } from 'vue'
 import type { TreeNodeData } from '../../types/tree';
 import {useStorageStore} from '../../composables/useStorage';
 import TreeNodeField from '../../components/fields/TreeNodeField.vue';
-import { findNodeById } from '@/utils/treeMapper'
 
 const storageStore = useStorageStore();
 
@@ -14,7 +13,7 @@ const props = defineProps({
   },
   dnd: {
     type: Boolean,
-    default: true,
+    default: false,
   },
 })
 
@@ -24,33 +23,10 @@ defineEmits<{
   (e: 'drop', node: TreeNodeData): void
 }>();
 
+
 computed(() => {
   return storageStore.findStorageById(props.node.id, storageStore.storage) != null;
-});
-
-const draggedNode = ref<TreeNodeData | null>(null);
-
-const onDragStart = (node: TreeNodeData) => {
-  if (!props.dnd) {
-    return;
-  }
-  emit('dragstart', draggedNode.value);
-  draggedNode.value = node;
-};
-
-const onDrop = (node: TreeNodeData) => {
-  if (!props.dnd || !draggedNode.value) {
-    return;
-  }
-  emit('drop', { draggedNode: draggedNode.value, targetNode: node });
-  draggedNode.value = null;
-};
-
-const allowDrop = (event: DragEvent) => {
-  if (props.dnd) {
-    event.preventDefault();
-  }
-};
+})
 
 // Transition Hooks
 const beforeEnter = (el: Element): void => {
@@ -82,12 +58,13 @@ const leave = (el: Element): void => {
 <template>
   <li
     :draggable="dnd"
-    @dragstart="onDragStart"
-    @dragover="allowDrop"
-    @drop="onDrop"
+    @dragstart="$emit('dragstart', node)"
+    @dragover="dnd && $event.preventDefault()"
+    @drop="$emit('drop', node)"
   >
-    <TreeNodeField :node="node" @toggle="$emit('toggle', $event)" />
-
+    <TreeNodeField
+      :node="node"
+      @toggle="$emit('toggle', $event)" />
     <transition
       name="expand-fade-y"
       @before-enter="beforeEnter"
@@ -105,7 +82,6 @@ const leave = (el: Element): void => {
           @drop="$emit('drop', $event)"
           @toggle="$emit('toggle', $event)"
         />
-
       </ul>
     </transition>
   </li>
